fix various thing in io, implement rom mask

+++ gba/include/gba/arm/arm7tdmi.h
@@ -205,17 +205,17 @@ private:
 
     [[nodiscard]] u32 read_16_signed(u32 addr, mem_access access) noexcept;
     [[nodiscard]] u32 read_16_aligned(u32 addr, mem_access access) noexcept;
-    [[nodiscard]] u32 read_16(u32 addr, mem_access access) noexcept;
+    [[nodiscard]] u16 read_16(u32 addr, mem_access access) noexcept;
     void write_16(u32 addr, u16 data, mem_access access) noexcept;
 
     [[nodiscard]] u32 read_8_signed(u32 addr, mem_access access) noexcept;
-    [[nodiscard]] u32 read_8(u32 addr, mem_access access) noexcept;
+    [[nodiscard]] u8 read_8(u32 addr, mem_access access) noexcept;
     void write_8(u32 addr, u8 data, mem_access access) noexcept;
 
     [[nodiscard]] u32 read_bios(u32 addr) noexcept;
     [[nodiscard]] u32 read_unused(u32 addr) noexcept;
 
-    [[nodiscard]] u32 read_io(u32 addr) noexcept;
+    [[nodiscard]] u8 read_io(u32 addr) noexcept;
     void write_io(u32 addr, u8 data) noexcept;
 
     void update_waitstate_table() noexcept;
